% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addPoints}
\alias{addPoints}
\title{Add points to an existing plot}
\usage{
addPoints(
  plt = last.plot(),
  x = NULL,
  y = NULL,
  color = "black",
  shape = 16,
  size = 1
)
}
\arguments{
\item{plt}{A \pkg{ggplot2} plot object.}

\item{x, y}{Coordinates of the points.}

\item{color}{Name of the color of the points.}

\item{shape}{Type of point, specified by either an integer (0-25) or a name, see \code{\link{vignette("ggplot2-specs")}}}

\item{size}{Size of the point in mm.}
}
\value{
A \pkg{ggplot2} object containing the updated plot.
}
\description{
Function to add points to an existing plot.
}
\examples{
library(minigst)

# Load Data and create Db
data("Scotland")
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))

# Create a plot of  the variable Elevation
plt=dbplot_point(db=db,size="Elevation")

# Add two triangular-shaped, orange points the plot
plt=addPoints(plt=plt,x=c(100,400),y=c(600,1100),color="orange", shape = 17, size = 3)

# Display the plot
print(plt)


}
