% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{addSel}
\alias{addSel}
\alias{clearSel}
\title{Add a selection to a Db}
\usage{
addSel(db, sel)

clearSel(db)
}
\arguments{
\item{db}{Db object.}

\item{sel}{A vector of boolean values indicating which points of \code{db} are selected. The length of \code{sel} must be equal to the number of points/rows in \code{db} (given by \code{nrow(db[])}).}
}
\value{
The function updates \code{db} and returns nothing.
}
\description{
Function to add a new selection/mask to an existing \pkg{gstlearn} Db object.
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db$display()

# Select points with an elevation greater than 100
selVar=(db["Elevation"]>100)

# Add selection to Db
addSel(db=db, sel = selVar)
db$display()

# Clear the selection from the Db
clearSel(db=db)
db$display()

}
