% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{addVarToDb}
\alias{addVarToDb}
\title{Add a variable to a Db}
\usage{
addVarToDb(db, var, vname)
}
\arguments{
\item{db}{Db object.}

\item{var}{A vector (or a matrix) containing the new variable(s) to be added to \code{db}. The length (or number of rows) of `var` must be equal to the number of points/rows in \code{db} (given by \code{nrow(db[])}).}

\item{vname}{Name(s) of the new variables, stored as a (vector of) string(s).}
}
\value{
The function updates \code{db} and returns nothing.
}
\description{
Function to add a new variable/column to an existing `gstlearn` Db object.
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db$display()

# Create 2 new variables in a matrix
newVars=matrix(rnorm(2*nrow(db[])),ncol=2)  # Matrix of normal values
head(newVars)

# Add variables to db
addVarToDb(db=db, var = newVars, vname = c("Newvar1", "Newvar2"))
db$display()

}
