% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{createModel}
\alias{createModel}
\title{Create a model}
\usage{
createModel(
  struct = "SPHERICAL",
  range = 0.3,
  sill = 1,
  param = 1,
  ndim = 2,
  mean = 0,
  isScale = FALSE
)
}
\arguments{
\item{struct}{Vector containing the names of the desired basic structures. The list of available structures is obtained by calling the function \code{\link{printAllStruct}}.}

\item{range}{Value or vector specifying the range (see `isScale` for interpretation) of each structure. Must be of size 1 (if all the structures share the same value) or have the same size as \code{struct}.}

\item{sill}{Value or vector specifying the sill/variance of each structure. Must be of size 1 (if all the structures share the same value) or have the same size as \code{struct}.}

\item{param}{Value or vector specifying the extra parameter (eg. smoothness parameter for Matérn covariances) of each structure. Must be of size 1 (if all the structures share the same value) or have the same size as \code{struct}.}

\item{ndim}{Space dimension of the model.}

\item{mean}{Mean of the model (0 by default).}

\item{isScale}{Whether the specified `range` correspond to a correlation length (`FALSE`, choice by default) or a scaling factor (`TRUE`).}
}
\value{
A \pkg{gstlearn} Model object.
}
\description{
Function to create a \pkg{gstlearn} Model object (corresponding to a single variable with an isotropic covariance function).
}
\examples{
library(minigst)

## Parameters of the model
struct_names = c("SPHERICAL", "K-BESSEL")
ranges=c(0.3,0.5)
variances=c(0.1,1)
params=c(1,1) # The first value will be ignored since the correponding covariance function ("SPHERICAL") does not require an extra parameter.

## Create model
model=createModel(struct=struct_names, range = ranges, sill = variances, param = params,ndim=2)

}
