% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dbplot_grid}
\alias{dbplot_grid}
\title{Plot variables in a Db}
\usage{
dbplot_grid(
  dbGrid,
  color = NULL,
  contour = NULL,
  colValLimits = NULL,
  cmap = "viridis",
  naColor = NA,
  colorLegendTitle = title,
  contourLegendTitle = NULL,
  asp = 0,
  xlab = NA,
  ylab = NA,
  title = NA,
  ...
)
}
\arguments{
\item{dbGrid}{DbGrid object.}

\item{color}{Name of the variable defining the color of the grid points.}

\item{contour}{Name of the variable defining the countours of the plot.}

\item{colValLimits}{Vector specifying the limits applied to the variable supplied in \code{color}.}

\item{cmap}{Color map used to plot the points by color. Can be specified by a name (cf. \code{\link{printAllPalettes}} to display vailable choices), or by a vector of color names which will be interpolated to create a palette.}

\item{naColor}{Color used for NA values.}

\item{colorLegendTitle}{Title of the color legend.}

\item{contourLegendTitle}{Title of the contour legend.}

\item{asp}{Aspect ratio of the plot: 0 = No scaling of the axes, 1 = Axes have the same scale.}

\item{xlab}{Title of the x-axis.}

\item{ylab}{Title of the y-axis.}

\item{title}{Title of plot.}

\item{...}{Additional arguments passed to the \pkg{ggplot2} function \code{geom_raster}.}
}
\value{
A \pkg{ggplot2} object containing the plot. If stored in a variable, use the function \code{print} to display the plot.
}
\description{
Function to plot variables in a Db object as isolated points. The value of the variables are used to define
the color, size and/or labels of the points.
}
\examples{
library(minigst)

# Load Data
data("ScotlandGrid")

# Create DbGrid
dbG=dfToDbGrid(df=ScotlandGrid,coordnames = c("Longitude","Latitude"))

# Plot the variables Elevation by color using the "viridis" palette
dbplot_grid(dbG,color="Elevation",cmap = "viridis")

# Store the plot, and then display it
plt= dbplot_grid(dbG,color="Elevation",cmap = "viridis",title="Printed plot")
print(plt)

}
