% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{.checkCovParam}
\alias{.checkCovParam}
\title{Function to check the covariance parameters proposed by the user}
\usage{
.checkCovParam(param, nameParam, n)
}
\description{
Function to check the covariance parameters proposed by the user
}
\keyword{internal}
