% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{.checkIfRegStep}
\alias{.checkIfRegStep}
\title{Function to check if a vector properly defines a set of grid coordinates,
i.e. if the values are ordered}
\usage{
.checkIfRegStep(yseq, tolRelError = 1e-06)
}
\arguments{
\item{yseq}{Vector}

\item{tolRelError}{Tolerance value to accept that two values are equal}
}
\description{
Function to check if a vector properly defines a set of grid coordinates,
i.e. if the values are ordered
}
\keyword{internal}
