% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{.checkVariableNames}
\alias{.checkVariableNames}
\title{Function to delete variable which already exists in a Db}
\usage{
.checkVariableNames(db, vname)
}
\description{
Function to delete variable which already exists in a Db
}
\keyword{internal}
