% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{.createneigh}
\alias{.createneigh}
\title{Function to create a gstlearn object containing the parameters of 
the neighborhood used to perform kriging}
\usage{
.createneigh(neigh)
}
\description{
Function to create a gstlearn object containing the parameters of 
the neighborhood used to perform kriging
}
\keyword{internal}
