% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{.createVarioParam}
\alias{.createVarioParam}
\title{Function to create a gstlearn object containing the parameters needed to compute
an experimental variogram.
See the documentation of the function \code{vario_exp}
for details about the parameters.}
\usage{
.createVarioParam(
  db,
  dir = NULL,
  nlag = 20,
  dlag = 100,
  toldis = 0.5,
  tolang = NULL
)
}
\description{
Function to create a gstlearn object containing the parameters needed to compute
an experimental variogram.
See the documentation of the function \code{vario_exp}
for details about the parameters.
}
\keyword{internal}
