% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.defineColour}
\alias{.defineColour}
\title{Functions to create color scales from palettes}
\usage{
.defineColour(
  palette,
  naColor = "transparent",
  flagDiscrete = FALSE,
  limits = NULL,
  title = NA
)
}
\description{
Functions to create color scales from palettes
}
\keyword{internal}
