% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{.deleteExistingVar}
\alias{.deleteExistingVar}
\title{Function to delete variable which already exists in a Db}
\usage{
.deleteExistingVar(db, vname)
}
\description{
Function to delete variable which already exists in a Db
}
