% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{.model_eval_base}
\alias{.model_eval_base}
\title{Function to evaluate a model between two points.
See \code{model_eval} for details on the parameters.}
\usage{
.model_eval_base(x, y, model, mode)
}
\description{
Function to evaluate a model between two points.
See \code{model_eval} for details on the parameters.
}
\keyword{internal}
