% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.plot_vario_base}
\alias{.plot_vario_base}
\title{Function to plot a single experimental variogram or cross-variogram and overlay a model on top.}
\usage{
.plot_vario_base(
  expvario = NA,
  model = NA,
  ivar = 0,
  jvar = 0,
  idir = -1,
  pairDisplay = NA,
  nlag = 100,
  hmax = NA,
  color = NULL,
  title = "",
  legend = TRUE,
  ...
)
}
\description{
See \code{plot_vario} for more info on the parameters.
}
\keyword{internal}
