% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{.rep_arg}
\alias{.rep_arg}
\title{Function to adapt the size of the model parameters passed to gstlearn functions
If n>0, it makes sure the argument arg has size 1 or n, and returns a copy of arg with size n.}
\usage{
.rep_arg(n, arg, argname)
}
\description{
Function to adapt the size of the model parameters passed to gstlearn functions
If n>0, it makes sure the argument arg has size 1 or n, and returns a copy of arg with size n.
}
\keyword{internal}
