% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{.resetNames}
\alias{.resetNames}
\title{Function to rename all the variables from a Db sharing the same locator}
\usage{
.resetNames(db, Loc, names)
}
\arguments{
\item{db}{Db object}

\item{Loc}{Locator object}

\item{names}{Vector of names (must be of the same as )}
}
\description{
Function to rename all the variables from a Db sharing the same locator
}
\keyword{internal}
