% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{getAllStruct}
\alias{getAllStruct}
\alias{printAllStruct}
\title{Basic structures for models}
\usage{
getAllStruct()

printAllStruct()
}
\value{
`getAllStruct` returns a dataframe (with two columns `Name` and `Description`) listing all the available basic structures, whereas `printAllStruct` only prints this dataframe and returns nothing.
}
\description{
Get the list of all available basic structures/covariance functions in \pkg{gstlearn}.
}
\examples{
library(minigst)

# Print list of basic structures
printAllStruct()

# Store the name of the 3rd covariance structure in the list ("SPHERICAL")
struct_name=getAllStruct()$Name[3]
print(struct_name)

}
