% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getSeq}
\alias{getSeq}
\title{Function to create breaks (used to plot size legend in `dbplot_point`)}
\usage{
getSeq(v, n, nd = 3)
}
\description{
Function to create breaks (used to plot size legend in `dbplot_point`)
}
\keyword{internal}
