% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussim.R
\name{kriging_mean}
\alias{kriging_mean}
\title{Compute the Kriging mean of a variable}
\usage{
kriging_mean(db, vname, model)
}
\arguments{
\item{db}{Db object containing the observations.}

\item{vname}{Name of the variable of interest.}

\item{model}{Model object giving the covariance/variogram model of the data.}
}
\value{
The function returns a numeric value giving the kriging mean of the variable of interest.
}
\description{
Function that calculates the kriging mean of a variable in a database.
}
\details{
...
}
\examples{
library(minigst)

# Database
data("Scotland")
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))

# Model
varioExp = vario_exp(db=db, vname="Elevation", dir=c(30,-30), nlag=20, dlag=10.)
model = model_fit(varioExp, struct=c("NUGGET","SPHERICAL", "SPHERICAL"), pruneModel = TRUE)

# Kriging mean
km = kriging_mean(db = db, vname = "Elevation", model = model)
km

}
