% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussim.R
\name{minirgrf}
\alias{minirgrf}
\title{Simulate a Gaussian random field with the turning bands algorithm}
\usage{
minirgrf(
  target,
  model,
  nsim,
  ntuba = 1000,
  seed = 43431,
  simname = NULL,
  dbcond = NULL,
  vcond = NULL
)
}
\arguments{
\item{target}{Db object containing the target points on which we want to simulate the model, and to which the result will be added.}

\item{model}{Model object defining the model we want to simulate.}

\item{nsim}{Number of samples to generate.}

\item{ntuba}{Number of turning bands to use (Default 100).}

\item{seed}{Optional integer specifying the seed of the algorithm for reproducibility purposes (defaults to 43431).}

\item{simname}{Optional name of the variable displaying the simulated values in the database.}

\item{dbcond}{Optional Db object containing the conditioning data.}

\item{vcond}{Optional character specifying which variable in \code{dbcond} is to be used as conditioning data.}
}
\value{
The function updates \code{target} and returns nothing.
}
\description{
Function that adds simulations of a Gaussian random field to a database that indicates the target locations.
}
\details{
...
}
\examples{
library(minigst)

# Target grid
tgt=createDbGrid(coords=list(seq(from=0,to=1,length.out=100),seq(from=0,to=1,length.out=100)),coordnames=c("xcoord", "ycoord"))

# Model
mod=createModel(struct=c("EXPONENTIAL", "NUGGET"), range = 0.3, sill = c(1,0.1), ndim=2, mean = 3)

# Simulation
minirgrf(target=tgt, model=mod, nsim = 1, simname = "Mysim") 
dbplot_grid(db=tgt,color="Mysim", cmap = "Spectral", title = "Simulation", colorLegendTitle= "Value")

}
