% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{model_covMat}
\alias{model_covMat}
\title{Compute a covariance or variogram matrix from a model}
\usage{
model_covMat(x, y = NULL, model = createModel(), mode = "COV")
}
\arguments{
\item{x, y}{Matrices (of same size) whose rows specify the pairs of points at which the covariance is evaluated. The number of columns of these matrices must be the same as the space dimension of the model.}

\item{model}{Model object to evaluate.}

\item{mode}{Whether to compute the variogram values (\code{mode="VG"}) or covariance values (\code{mode="VG"}).}
}
\value{
A matrix of size \code{nrow(x)}*\code{nrow(y)} containing the covariance/variogram between each point in \code{x} and each point in \code{y}.
}
\description{
Function to compute a covariance or variogram matrix between pairs of points, from a \pkg{gstlearn} Model object.
}
\details{
If the argument \code{y} is set to \code{NULL} (Default), then \code{y} is set to \code{x}.
}
\examples{
library(minigst)

## Create model
model=createModel(struct=c("EXPONENTIAL", "NUGGET"), range = 0.3, sill = c(1,0.1), ndim=2)

## Evaluate the covriance matrix on some random points in the unit square
xx=matrix(runif(20),nrow=10,ncol=2)
yy=matrix(runif(10),nrow=5,ncol=2)
covMat=model_covMat(x=xx,y=yy,model=model,mode="COV")



}
