% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot correlation plot of a couple of variables in a gstlearn db}
\usage{
plot_correlation(
  db,
  vnamex,
  vnamey,
  reg = T,
  sameaxes = T,
  nbin = 100,
  title,
  xlab,
  ylab
)
}
\arguments{
\item{db}{Db object.}

\item{vnamex}{name of the variable to be represented in abscissa}

\item{vnamey}{name of the variable to be represented in ordinate}

\item{reg}{Boolean, should the regression line be plotted}

\item{sameaxes}{Boolean, should both variables be represented on the same axes}

\item{title}{Title of plot.}

\item{xlab}{label of the abscissa}

\item{ylab}{label of the ordinate}
}
\value{
A \pkg{ggplot2} object containing the plot. If stored in a variable, use the function \code{print} to display the plot.
}
\description{
Function to plot the histogram of a variable in a gstlearn db
}
\examples{
library(minigst)

#todo

}
