% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot correlation plot of a couple of variables in a gstlearn db}
\usage{
plot_scatter(
  db,
  vnamex,
  vnamey,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  asPoints = FALSE,
  nbins = 25,
  cmap = "viridis",
  regrLine = FALSE,
  regrLineColor = "blue",
  regrLineType = "solid",
  regrLineWidth = 1.25,
  eqLine = FALSE,
  eqLineColor = "red",
  eqLineType = "dashed",
  eqLineWidth = 1.25,
  verbose = TRUE
)
}
\arguments{
\item{db}{Db object.}

\item{vnamex}{Name of the variable to be represented in abscissa}

\item{vnamey}{Name of the variable to be represented in ordinate}

\item{xlab}{Label of x-axis}

\item{ylab}{Label of y-axis}

\item{title}{Title of plot.}

\item{asPoints}{Whether to plot points using indivudal symbols or counting bins}

\item{nbins}{Number of discretization bins}

\item{cmap}{Color map used to plot the points by color. Can be specified by a name (cf. \code{\link{printAllPalettes}} to display available choices), or by a vector of color names which will be interpolated to create a palette.}

\item{regrLine}{Whether to add a regression line to the plot}

\item{regrLineColor, regrLineType, regrLineWidth}{Color, type and width of the regression line}

\item{eqLine}{Whether to add a regression line x=y to the plot}

\item{eqLineColor, eqLineType, eqLineWidth}{Color, type and width of the  line x=y}

\item{verbose}{Whether to print the regression results (when `regrLine=TRUE`)}
}
\value{
A \pkg{ggplot2} object containing the plot. If stored in a variable, use the function \code{print} to display the plot.
}
\description{
Function to plot the histogram of a variable in a gstlearn db
}
