% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{printAllPalettes}
\alias{printAllPalettes}
\title{Color palettes}
\usage{
printAllPalettes()
}
\value{
Prints the list of color palette names and returns nothing.
}
\description{
Print the list of predefined color palettes in \pkg{gstlearn}.
}
\examples{
library(minigst)

# Print list of basic structures
printAllPalettes()

}
