% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{printAllStruct}
\alias{printAllStruct}
\title{Basic structures for models}
\usage{
printAllStruct()
}
\value{
Prints the list of names and returns nothing.
}
\description{
Print the list of all available basic structures/covariance functions in \pkg{gstlearn}.
}
\examples{
library(minigst)

# Print list of basic structures
printAllStruct()

}
