% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussim.R
\name{setMean}
\alias{setMean}
\title{Set the mean in a model}
\usage{
setMean(model, mu)
}
\arguments{
\item{model}{Model object.}

\item{mu}{Value of the mean.}
}
\value{
The function updates \code{model} and returns nothing.
}
\description{
Function that sets the value of the mean in a Model object.
}
\details{
...
}
\examples{
library(minigst)

# Database
data("Scotland")
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))

# Model
varioExp = vario_exp(db=db, vname="Elevation", dir=c(30,-30), nlag=20, dlag=10.)
model = model_fit(varioExp, struct=c("NUGGET","SPHERICAL", "SPHERICAL"), pruneModel = TRUE)

# Kriging mean
km = kriging_mean(db = db, vname = "Elevation", model = model)

# Set the mean in the model
setMean(model,km)

}
