% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{summaryStats}
\alias{summaryStats}
\title{Compute summary statistics}
\usage{
summaryStats(db, vname, stat = NULL, onlyCommon = FALSE)
}
\arguments{
\item{db}{Db object.}

\item{vname}{Name(s) of the variable(s) for which statistics are desired, stored as a (vector of) string(s).}

\item{stat}{Vector containing the names of the statistics to compute. Available choices are listed the section \emph{Details} below. The default value (\code{NULL}) computes all the available statistics.}

\item{onlyCommon}{Compute the statistics based only on the points for which all variables are defined (i.e. are not \code{NA}).}
}
\value{
A dataframe containing the selected statistics computed for each specified variable.
}
\description{
Function to compute summary statistics on the variables of a Db.
}
\details{
The available choices for the statistics are as follows:
\describe{
   \item{\code{"NUM"}}{Number of defined values}
   \item{\code{"MIN"}}{Minimum}
   \item{\code{"MAX"}}{Maximum}
   \item{\code{"MEAN"}}{Mean}
   \item{\code{"STDV"}}{Standard deviation}
   \item{\code{"MED"}}{Median}
   }
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))

# Compute summary statistics
sumStats=summaryStats(db = db, vname = c("Elevation","January_temp"))
print(sumStats)

}
