% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{vario_cloud}
\alias{vario_cloud}
\title{Compute a variogram cloud}
\usage{
vario_cloud(db, vname, dir = NULL, gridRes = 100, tolang = 22.5, plot = TRUE)
}
\arguments{
\item{db}{Db object.}

\item{vname}{Name of the variable.}

\item{dir}{Direction used to compute the variogram. Can be either \code{NULL} (Default) for an omnidirectional variogram cloud, or a single direction specified either by an angle in degrees (only in 2D and the value denote the angle between the horizontal axis and the desired direction), or as a matrix with a single row containing a direction vector.}

\item{gridRes}{Number or vector of size 2 specifying the resolution of the computation grid. See \emph{Details}.}

\item{tolang}{Tolerance on the angles. See \emph{Details}.}

\item{plot}{Whether or not to plot the variogram cloud.}
}
\value{
A DbGrid containing the computed variogram cloud.
}
\description{
Function to compute a variogram cloud for a variable in a `gstlearn` Db object.
}
\details{
The variogram cloud is computed as a 2D grid with dimensions given by \code{gridRes}. The first axis discretizes the set of possible distances between pairs of points in \code{db}.
The second axis discretizes the set of possible variogram values between between pairs of points in \code{db}. The value computed at a given grid node corresponds to the number of pairs of points in \code{db} falling in the corresponding grid cell.
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db$display()

# Compute and plot the omnidirectional variogram cloud of the variable "Elevation".
varioCloudGrid = vario_cloud(db=db, vname="Elevation", plot = TRUE)

}
