% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{vario_exp}
\alias{vario_exp}
\title{Compute an experimental variogram}
\usage{
vario_exp(
  db,
  vname,
  dir = NULL,
  nlag = 20,
  dlag = 100,
  toldis = 0.5,
  tolang = 22.5
)
}
\arguments{
\item{db}{Db object.}

\item{vname}{Name(s) of the variable(s), stored as a (vector of) string(s).}

\item{dir}{Direction used to compute the variogram. The default (\code{NULL}) reverts to an omnidirectional variogram. The directions can be specified either by a vector of angles in degrees (only in 2D and the value(s) denote the angle between the horizontal axis and the desired direction), or as a matrix whose rows contain direction vectors.}

\item{nlag}{Number of lags used to compute the variograms.}

\item{dlag}{Distance of the lags.}

\item{toldis}{Tolerance on the distance (between 0 and 1). See \emph{Details}.}

\item{tolang}{Tolerance on the angles (between 0 and 90). See \emph{Details}.}
}
\value{
The function returns a \pkg{gstlearn} object containing the experimental variogram(s) (and cross-variograms).
}
\description{
Function to compute an experimental variogram for a single variable, or experimental variogram and cross-variograms for a set of variables.
}
\details{
The parameter \code{toldis} represents a tolerance on the distance when computing the variogram value, and is expressed as proportion.
For  a given lag value \eqn{h}, all the pairs separated by distance between (\code{h}-\code{toldis}*\code{dlag}) and (\code{h}+\code{toldis}*\code{dlag})
are used to compute the correponding variogram value \eqn{gamma(h)}.

The parameter \code{tolang} represents a tolerance on the angle and is expressed as an angle.
For a given lag value \eqn{h} and a given angle \eqn{theta}, all the pairs with an angle between (\eqn{theta}-\code{tolang}) and (\eqn{theta}+\code{tolang})
are used to compute the correponding variogram value \eqn{gamma(h)}.
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db$display()

# Create directional experimental variograms along the directions 30 deg and -30 deg for the variable "Elevation".
varioExpDir = vario_exp(db=db, vname="Elevation", dir=c(30,-30), nlag=20, dlag=10.)
# varioExpDir$display() # Run to display the content of the experimental variogram object.
# Plot experimental variogram
plot_vario(varioExpDir, pairDisplay="size", title="Experimental variograms for the variable 'Elevation'")

# Create directional experimental (cross-)variograms along the directions 30 deg and -30 deg for the variables "Elevation" and "January_temp".
varioExpDir2var = vario_exp(db=db, vname=c("Elevation","January_temp"), dir=c(30,-30), nlag=20, dlag=10.)
# varioExpDir2var$display() # Run to display the content of the experimental variogram object.
# Plot experimental variogram
dev.new() # open new window
plot_vario(varioExpDir2var, title="Experimental (cross)-variograms for the variables 'Elevation' and 'January_temp")


}
