% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vario.R
\name{vario_map}
\alias{vario_map}
\title{Compute a variogram map}
\usage{
vario_map(db, vname, gridRes = 20, plot = T)
}
\arguments{
\item{db}{Db object.}

\item{vname}{Name of the variable.}

\item{gridRes}{Number specifying the resolution of the computation grid. See \emph{Details}.}

\item{plot}{Whether or not to plot the variogram cloud.}
}
\value{
A DbGrid containing the computed variogram map.
}
\description{
Function to compute a variogram map for a variable in a `gstlearn` Db object.
}
\details{
The variogram map is computed on a 2D grid, centered at the origin, and with dimensions (2*\code{gridSize}+1) * (2*\code{gridSize}+1).
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db$display()

# Compute and plot the omnidirectional variogram cloud of the variable "Elevation".
varioMapGrid = vario_map(db=db, vname="Elevation", plot = TRUE)

}
