% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{xvalidation}
\alias{xvalidation}
\title{Compute the cross validation error}
\usage{
xvalidation(
  dbin,
  vname,
  drift = NULL,
  model,
  type = "ordinary",
  prefix = "OK",
  mean = NULL,
  neighborhood = "unique",
  std = TRUE
)
}
\arguments{
\item{dbin}{Db containing the data.}

\item{vname}{Name of the variable, stored as a string.}

\item{model}{A \pkg{gstlearn} Model object containing a fitted model.}

\item{type}{Type of kriging, either "simple" or "ordinary".}

\item{mean}{Mean value if simple kriging is requested}

\item{neighborhood}{Type of neighborhood to used, either "unique" or a vector of the form (radius), (nmin, nmax) or (nmin, nmax, radius) see details.}

\item{std}{Boolean value indicating whether the prediction standard deviation should be computed.}
}
\value{
a db containing the same variables as the input db + two new variables: 
One variable (with name ending with .esterr) containing the cross-validation errors, 
and one variable (with name ending with .stderr) containing the cross-validation errors normalized by the kriging standard deviation.
}
\description{
Function to perform the (1-fold) cross validation for a given kriging exercise.
}
\details{
For each observation location x_i, compute the kriging prediction at x_i when using all the observations except the one located at x_i. 
The cross-validation error at x_i is then defined as the difference between this prediction and the actual observed value at x_i.
}
\examples{
library(minigst)

#todo


 

}
