% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dbplot_grid}
\alias{dbplot_grid}
\title{Plot variables in a Db}
\usage{
dbplot_grid(
  db,
  color = NULL,
  cat_color = NULL,
  contour = NULL,
  cmap = NULL,
  colValLimits = NULL,
  naColor = "transparent",
  nLevels = NULL,
  hideLegend = FALSE,
  legendTitle = NA,
  asp = 1,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  add = FALSE
)
}
\arguments{
\item{color, cat_color}{Name of the variable defining the color of the points: use `color` for continuous variables and `cat_color` for categorical variables.}

\item{contour}{Name of the variable defining the countours of the plot.}

\item{cmap}{Color map used to plot the points by color. Can be specified by a name (cf. \code{\link{printAllPalettes}} to display vailable choices), or by a vector of color names which will be interpolated to create a palette.}

\item{colValLimits}{Range of values for teh color mapping.}

\item{naColor}{Color used for NA values.}

\item{nLevels}{Number of levels for the contour plot.}

\item{hideLegend}{Whether to hide the legend.}

\item{legendTitle}{Title of the legend.}

\item{asp}{Aspect ratio of the plot: 0 = No scaling of the axes, 1 = Axes have the same scale.}

\item{xlab}{Title of the x-axis.}

\item{ylab}{Title of the y-axis.}

\item{title}{Title of plot.}

\item{add}{Either a boolean or \pkg{ggplot2} object. If \code{FALSE} (Default), then a new plot is created. If \code{TRUE}, then the points are added to the last plot (see \pkg{ggplot2::last_plot()}) specifying if the points are to be added to the last plot or not}

\item{dbGrid}{DbGrid object.}
}
\value{
A \pkg{ggplot2} object containing the plot. If stored in a variable, use the function \code{print} to display the plot.
}
\description{
Function to plot variables in a Db object as isolated points. The value of the variables are used to define
the color, size and/or labels of the points.
}
\details{
Note that you can specify only one of the arguments `color`, `cat_color` and `contour`.
}
\examples{
library(minigst)

# Load Data
data("ScotlandGrid")

# Create DbGrid
dbG=dfToDbGrid(df=ScotlandGrid,coordnames = c("Longitude","Latitude"))

# Plot the variables Elevation by color using the "viridis" palette
dbplot_grid(dbG,color="Elevation",cmap = "viridis")

# Store the plot, and then display it
plt= dbplot_grid(dbG,color="Elevation",cmap = "viridis",title="Printed plot")
print(plt)

}
