% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dbplot_point}
\alias{dbplot_point}
\title{Plot variables in a Db}
\usage{
dbplot_point(
  db,
  color = NULL,
  size = NULL,
  label = NULL,
  sizeRange = c(1, 5),
  absSize = FALSE,
  sizeLegendTitle = NULL,
  cmap = NULL,
  discreteVal = FALSE,
  colorLegendTitle = NULL,
  labelColor = "black",
  pointColor = "black",
  labelLegendTitle = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{db}{Db object.}

\item{color}{Name of the variable defining the color of the points.}

\item{size}{Name of the variable defining the size of the points.}

\item{label}{Name of the variable defining the labels of the points.}

\item{sizeRange}{Vector specifying the range of sizes when plotting the points by size.}

\item{absSize}{Whether the absolute value of the variable should be used to define the size of the points.}

\item{sizeLegendTitle}{Title of the size legend.}

\item{cmap}{Color map used to plot the points by color. Can be specified by a name (cf. \code{\link{printAllPalettes}} to display vailable choices), or by a vector of color names which will be interpolated to create a palette.}

\item{discreteVal}{Whether the variable defining the color is discrete or not.}

\item{colorLegendTitle}{Title of the color legend.}

\item{labelColor}{Name of the color used to write labels.}

\item{pointColor}{Default color of points when \code{color} is not specified.}

\item{labelLegendTitle}{Title of the label legend.}

\item{xlab}{Title of the x-axis.}

\item{ylab}{Title of the y-axis.}

\item{title}{Title of plot.}

\item{add}{Either a boolean or \pkg{ggplot2} object. If \code{FALSE} (Default), then a new plot is created. If \code{TRUE}, then the points are added to the last plot (see \pkg{ggplot2::last_plot()}) specifying if the points are to be added to the last plot or not}

\item{...}{Additional arguments passed to the \pkg{ggplot2} functions \code{geom_point} and \code{geom_text}.}
}
\value{
A \pkg{ggplot2} object containing the plot. If stored in a variable, use the function \code{print} to display the plot.
}
\description{
Function to plot variables in a Db object as isolated points. The value of the variables are used to define
the color, size and/or labels of the points.
}
\examples{
library(minigst)

# Load Data
data("Scotland")

# Create Db
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))

# Plot the variables Elevation (for the size of the points) and Longitude (for the color of the points)
dbplot_point(db=db,size="Elevation",color="Longitude")

# Store the plot, and then display it
plt=dbplot_point(db=db,size="Elevation",title="Printed plot")
print(plt)

}
