% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Db.R
\name{dfToDb}
\alias{dfToDb}
\alias{dfToDbGrid}
\title{Create a Db or a DbGrid from a dataframe}
\usage{
dfToDb(df, coordnames, isGrid = FALSE)

dfToDbGrid(df, coordnames)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{coordnames}{A vector containing the names of the columns of \code{df} that define the spatial coordinates of the data.}

\item{isGrid}{Whether the data in \code{df} should be loaded as a set of isolated points (i.e. a Db) or as a grid (i.e. a DbGrid).}
}
\value{
A Db object.
}
\description{
Function to convert a dataframe into a \pkg{gstlearn} Db or DbGrid object.
}
\details{
When loading grid data in \code{df}, if the data points are a subset of a grid, a boolean variable \code{Sel} is added to the DbGrid to indicate if each point belong to the initial dataset

Calling the function \code{dfToDbGrid} is the same as calling the function \code{dfToDb} with the option \code{isGrid=TRUE}.
}
\examples{
library(minigst)

# Load Data
data("Scotland")
print(head(Scotland))

# Create Db of isolated points from dataframe
db=dfToDb(df=Scotland, coordnames=c("Longitude", "Latitude"))
db

# Load Grid data
data("ScotlandGrid")
print(head(ScotlandGrid))

# Create DbGrid from dataframe
dbG=dfToDbGrid(df=ScotlandGrid,coordnames=c("Longitude","Latitude"))
dbG

}
